Fix schema prefixing problem in the ALTER TABLE statements
Fixes #565

With this change we prefix all table names in the ALTER TABLE statements before sending
them to worker nodes. Query deparsing is not available for ALTER statements at the moment
therefore it is not possible to change Query tree, deparse it and send resulting query to
worker. As workaround, we use string manipulation to prefix table names. An alternative
is changing worker_apply_shard_ddl_command function to accept schema name in addition to
shard id. It is open for discussion to decide how we handle this situation.

we use string man