Add concept of RTE identifiers.
For the improved router pushdown we need to track RTEs through postgres'
planning, so we can update them to point to remote nodes
instead.

As there's no nice builtin way to do so, add IdentifyRTE() and
GetRTEIdentity(). These are implemented somewhat hackishly, but since
that's mostly hidden, it shouldn't be too bad.